requirements.txt
pandas
requests
ping3
python-dotenv
openpyxl
